code/src/tools/geou-tsapp-0001/requirements.txt
oss2
PyYAML
requests==2.23.0
scipy
scikit-image
python-dotenv
whatimage
matplotlib
opencv-python==4.11.0.86
pillow
minio
flasgger
colorlog
apscheduler
ultralytics==8.3.128 
ultralytics-thop==2.0.14
pydantic
fastapi
uvicorn
httpx
pyheif